refactored graph.js to multiple files; formatted according to eslint
