+++ claimarg_prototype/core/templates/post_and_show_messages.html
@@ -6,7 +6,7 @@
     <link rel="shortcut icon" type="image/x-icon" href="{% static 'images/favicon.ico' %}">
     <link rel="stylesheet" href="{% static 'css/styles.css' %}">
     <script src="https://unpkg.com/vis-network/standalone/umd/vis-network.min.js"></script>
-    <script src="{% static 'js/graph.js' %}?v=1.0.2"></script>
+    <script type="module" src="{% static 'js/event-handlers.js' %}?v=0.0.0"></script>
 </head>
 <body>
     {% include 'partials/account_links.html' %}
